% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_risk.R
\name{clinic_risk}
\alias{clinic_risk}
\title{cbind clinical data and riskdata}
\usage{
clinic_risk(immue_RNA, riskdata)
}
\arguments{
\item{immue_RNA}{immue_RNA list data}

\item{riskdata}{riskdata dataframe}
}
\description{
cbind clinical data and riskdata
}
