% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2df.R
\name{count2df}
\alias{count2df}
\title{clear count data and clinical data}
\usage{
count2df(
  count_cart,
  count_metadata,
  clinical_cart,
  pattern = "htseq.counts.gz"
)
}
\arguments{
\item{count_cart}{directory of count cart}

\item{count_metadata}{directory of count metadata}

\item{clinical_cart}{directory of clinical}

\item{pattern}{file patern defalut is htseq.counts.gz}
}
\value{
mRNA, lncRNA, clinical data
}
\description{
clear count data and clinical data
}
