% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_edgeR_mRNA.R
\name{DEG_edgeR_mRNA}
\alias{DEG_edgeR_mRNA}
\title{DEG analysis using edgeR package for mRNA}
\usage{
DEG_edgeR_mRNA(d)
}
\arguments{
\item{d}{results of FPKM2df()}
}
\description{
DEG analysis using edgeR package for mRNA
}
