% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt1.R
\name{filt1}
\alias{filt1}
\title{Filt subject by minium time}
\usage{
filt1(d, minTime = 30)
}
\arguments{
\item{d}{result of FPKM2df()}

\item{minTime}{default is 30}
}
\description{
Filt subject by minium time
}
