% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt2.R
\name{filt2}
\alias{filt2}
\title{Filt gene expression}
\usage{
filt2(
  d,
  RNA = "mRNA",
  minExpPatientRatio = 0.4,
  minExpmean = 0.5,
  minExpsd = 0.5
)
}
\arguments{
\item{minExpPatientRatio}{dafault is 0.4. patitent ratio of minium expression of pratient}

\item{minExpmean}{dafault is 0.5, minium mean gene expression}

\item{minExpsd}{dafault is 0.5, minium sd gene expression}

\item{df}{dfframe with subject as rows and genes as columns}
}
\description{
Filt gene expression
}
