% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_cor.R
\name{filt_cor}
\alias{filt_cor}
\title{filt cor_RNA results}
\usage{
filt_cor(cor, minCor = 0.4, pvalue.cutoff = 0.05)
}
\arguments{
\item{cor}{results of cor_RNA()}

\item{minCor}{min correlation, default is 0.4}

\item{pvalue.cutoff}{pvalue cutoff ,default is 0.05}
}
\description{
filt cor_RNA results
}
