% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_deg.R
\name{filt_deg}
\alias{filt_deg}
\title{Filt deg by pvalue and lfc}
\usage{
filt_deg(deg, pvalue.cutoff = 0.01, lfc = 2)
}
\arguments{
\item{deg}{deg results}

\item{pvalue.cutoff}{0.05 default}

\item{lfc}{2 default}
}
\description{
Filt deg by pvalue and lfc
}
