% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_lncRNA.R
\name{filt_lncRNA}
\alias{filt_lncRNA}
\title{Filt lncRNA_surv from lncRNA_surv and cor_RNA}
\usage{
filt_lncRNA(lncRNA_surv, cor)
}
\arguments{
\item{lncRNA_surv}{lncRNA_surv}

\item{cor}{results of cor_RNA, should after filt_cor()}
}
\description{
Filt lncRNA_surv from lncRNA_surv and cor_RNA
}
