% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPKM2df.R
\name{FPKM2df}
\alias{FPKM2df}
\title{clear fpkm data and clinical data}
\usage{
FPKM2df(clinical_cart, trans_cart, trans_metadata, pattern = "FPKM.txt")
}
\arguments{
\item{clinical_cart}{directory of clinical}

\item{trans_cart}{directory of fpkm cart}

\item{trans_metadata}{directory of fpkm metadata}

\item{pattern}{file patern defalut is FPKM.txt}
}
\value{
mRNA, lncRNA, clinical data
}
\description{
clear fpkm data and clinical data
}
