% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{Volcano plot by ggplot2}
\usage{
ggvolcano(
  data,
  logFC = "logFC",
  pvalue = "PValue",
  symbol = NULL,
  pvalue.cutoff = 0.01,
  lgf = 2,
  alpha = 0.5,
  size = 2,
  colours = c("red", "blue", "gray"),
  size.border = 1,
  size.vhline = 0.75,
  size.axis.text = 13,
  size.axis.title = 14,
  size.symbol = 4,
  ticks.length = 0.15,
  ticks.width = 1,
  legend.text = c("Up", "Down", "NS"),
  legend.title = NULL,
  legend.position = "right",
  size.legend.text = 12,
  size.legend.title = 15,
  symbol.pvalue.cutoff = NULL,
  symbol.lgf = NULL,
  cat = TRUE,
  flip = FALSE,
  grid = FALSE
)
}
\arguments{
\item{data}{results of difference expression genes, which must contains
logFC and p value.}

\item{logFC}{default is logFC}

\item{pvalue}{default if PValue}

\item{symbol}{default if NULL}

\item{pvalue.cutoff}{default is 0.01}

\item{lgf}{default is 2}

\item{alpha}{for points, default is 0.5}

\item{size}{for points, default is 2}

\item{colours}{for points, default is red, bule and gray}

\item{size.border}{1}

\item{size.vhline}{0.75}

\item{size.axis.text}{13}

\item{size.axis.title}{14}

\item{size.symbol}{4}

\item{ticks.length}{0.15}

\item{ticks.width}{1}

\item{legend.text}{text for three groups, default is Up, Down and NS. NS means not significant}

\item{legend.title}{NULL}

\item{legend.position}{legend position, which can be right, left, top, bottom, or two of them. Or two numbers, eg: c(1,1)}

\item{size.legend.text}{12}

\item{size.legend.title}{15}

\item{symbol.pvalue.cutoff}{default is NULL, which takes value of pvalue.cutoff}

\item{symbol.lgf}{default is NULL, which takes value of lgf}

\item{cat}{logical, whether to output gene number for three classifications.}

\item{flip}{logical, default is TRUE, whether to flip coordinate}

\item{grid}{logical, default is FALSE, whether to show panel grid}
}
\value{
one ggplot2 volcano picture
}
\description{
Volcano plot by ggplot2
}
