% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_cox.R
\name{mv_cox}
\alias{mv_cox}
\title{multi-variable for cox regression}
\usage{
mv_cox(data, time, status, x, direction = "no", ...)
}
\arguments{
\item{data}{data}

\item{time}{time variable}

\item{status}{status variable}

\item{x}{x variable}

\item{direction}{should be one of no, both, backward, forward}

\item{...}{other arg}
}
\description{
multi-variable for cox regression
}
