% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TCGA_Clinical.R
\name{read_TCGA_Clinical}
\alias{read_TCGA_Clinical}
\title{Read Clinical Data of TCGA}
\usage{
read_TCGA_Clinical(Clinical_dir)
}
\arguments{
\item{Clinical_dir}{directory of TCGA clinical data, which contains one or more xml files}
}
\value{
one dataframe
}
\description{
Read Clinical Data of TCGA
}
